+++ CHANGELOG.md
@@ -17,6 +17,11 @@
     Click to see more.
   </summary>
 
+## :house: Internal
+
+- `collapsible-list`
+  - [#4276](https://github.com/wix/ricos/pull/4276) toolbar buttons and position
+
 </details>
 
 ## 8.72.15 (Jul 31, 2022)